pyproject.toml
[tool.poetry]
name = "github-resolver"
version = "0.1.0"
description = "OpenHands Issue Resolver"
authors = ["All Hands AI"]
license = "Proprietary"
readme = "README.md"
repository = "https://github.com/All-Hands-AI/github-resolver"
include = ["poetry.lock"]

[tool.poetry.dependencies]
python = "^3.11"
openhands-ai = { git = "https://github.com/All-Hands-AI/openhands.git", rev = "main" }
whatthepatch = "*"
pygithub = "^2.4.0"

[tool.poetry.group.dev.dependencies]
mypy = "*"
ruff = "*"
types-requests = "*"
types-whatthepatch = "*"

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-asyncio = "*"

